pyproject.toml
[tool.poetry]
name = "moneyforward-checker"
version = "0.1.0"
description = ""
authors = ["roto5296 <hikun5296@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
mfscraping = {git = "https://github.com/roto5296/moneyforward-scraping-py.git", branch = "master"}
mfscraping-asyncio = {git = "https://github.com/roto5296/moneyforward-scraping-asyncio-py.git", branch = "main"}
gspread-asyncio = "^1.9.0"
google-auth = "^2.22.0"

[tool.poetry.dev-dependencies]
flake8 = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"